* New command help, some bugs fixed

+++ src/cz/xds/command/delete.java
@@ -2,22 +2,26 @@ package cz.xds.command;
 
 import cz.xds.Command;
 import cz.xds.FileSystem;
-import cz.xds.FileSystemItem;
 import cz.xds.FileSystemException;
+import cz.xds.FileSystemItem;
 
-import java.util.Iterator;
 import java.io.PrintStream;
+import java.util.Iterator;
 
 /**
  * Vymaze polozku
  */
 public class delete implements Command {
     public void execute(FileSystem fs, PrintStream outStream, Object[] param) throws FileSystemException {
-        String name = (String)param[1];
+        if (param.length < 2) {
+            throw new FileSystemException(new String("Invalid usage. Use : " + help(true)));
+        }
+
+        String name = (String) param[1];
         Iterator it = fs.getCurrentDirectory().getIterator();
 
         while (it.hasNext()) {
-            FileSystemItem fsi = (FileSystemItem)it.next();
+            FileSystemItem fsi = (FileSystemItem) it.next();
             if (fsi.getName().equals(name)) {
                 fsi.delete();
                 return;
@@ -27,7 +31,7 @@ public class delete implements Command {
     }
 
     public String help(boolean briefOnly) {
-        return new String("delete");
-        //TODO: kompletni napoveda
+        return new String(briefOnly ? "delete itemname" : "delete - remove item from xds\n" +
+                "\titemname - item name. May be link, file or directory (which is deleted recursively)");
     }
 }
